+++ src/PromocaoHumana.Web/PromocaoHumana.Web/Data/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -221,9 +221,9 @@ namespace PromocaoHumana.Web.Data.Migrations
 
             modelBuilder.Entity("PromocaoHumana.Web.Domain.Doacao", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<DateTime>("DataRetirada")
@@ -234,11 +234,11 @@ namespace PromocaoHumana.Web.Data.Migrations
                         .HasColumnType("varchar")
                         .HasMaxLength(150);
 
-                    b.Property<long?>("FamiliaId")
-                        .HasColumnType("bigint");
+                    b.Property<int?>("FamiliaId")
+                        .HasColumnType("int");
 
-                    b.Property<long?>("LocalRetiradaId")
-                        .HasColumnType("bigint");
+                    b.Property<int?>("LocalRetiradaId")
+                        .HasColumnType("int");
 
                     b.Property<string>("QuemRetirou")
                         .IsRequired()
@@ -246,7 +246,7 @@ namespace PromocaoHumana.Web.Data.Migrations
                         .HasMaxLength(150);
 
                     b.Property<int>("Tipo")
-                        .HasColumnType("long");
+                        .HasColumnType("int");
 
                     b.HasKey("Id");
 
@@ -259,9 +259,9 @@ namespace PromocaoHumana.Web.Data.Migrations
 
             modelBuilder.Entity("PromocaoHumana.Web.Domain.Endereco", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<string>("Bairro")
@@ -304,9 +304,9 @@ namespace PromocaoHumana.Web.Data.Migrations
 
             modelBuilder.Entity("PromocaoHumana.Web.Domain.Familia", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<bool>("Ativa")
@@ -320,8 +320,8 @@ namespace PromocaoHumana.Web.Data.Migrations
                     b.Property<DateTime>("DataCadastro")
                         .HasColumnType("datetime");
 
-                    b.Property<long?>("EnderecoId")
-                        .HasColumnType("bigint");
+                    b.Property<int?>("EnderecoId")
+                        .HasColumnType("int");
 
                     b.Property<string>("NomeConjuge")
                         .HasColumnType("varchar")
@@ -332,8 +332,8 @@ namespace PromocaoHumana.Web.Data.Migrations
                         .HasColumnType("varchar")
                         .HasMaxLength(150);
 
-                    b.Property<long?>("ParoquiaId")
-                        .HasColumnType("bigint");
+                    b.Property<int?>("ParoquiaId")
+                        .HasColumnType("int");
 
                     b.Property<int>("QuantidadeFilhos")
                         .ValueGeneratedOnAdd()
@@ -351,9 +351,9 @@ namespace PromocaoHumana.Web.Data.Migrations
 
             modelBuilder.Entity("PromocaoHumana.Web.Domain.Igreja", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<string>("Cnpj")
@@ -361,8 +361,8 @@ namespace PromocaoHumana.Web.Data.Migrations
                         .HasColumnType("nvarchar")
                         .HasMaxLength(14);
 
-                    b.Property<long?>("EnderecoId")
-                        .HasColumnType("bigint");
+                    b.Property<int?>("EnderecoId")
+                        .HasColumnType("int");
 
                     b.Property<string>("Nome")
                         .IsRequired()
